requirements.txt
aiofiles==24.1.0
jsonargparse==4.35.0
langchain==0.3.19
langchain_anthropic==0.3.8
langchain_chroma==0.2.2
langchain_cohere==0.4.2
langchain_community==0.3.18
langchain_core==0.3.40
langchain_openai==0.3.7
langchain_text_splitters==0.3.6
langgraph==0.3.2
matplotlib==3.10.1
numpy==2.2.3
pandas==2.2.3
pydantic==2.10.6
python-dotenv==1.0.1
PyYAML==6.0.2
PyYAML==6.0.2
seaborn==0.13.2
setuptools==75.1.0
tiktoken==0.8.0
tqdm==4.67.1
